requirements.txt
Django
djangorestframework
django-cors-headers
pillow
gunicorn